# devdocs/openapi.yaml
openapi: 3.0.3
info:
  title: PostWEB 校园社区 接口文档
  description: 由 docs/api-protocol.md 编译。导入 Apifox 后会自动生成全部接口。
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [认证]
      summary: 注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: scu123 }
                password: { type: string, example: "123456" }
                nickname: { type: string, example: 东门猛男 }
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenUser' }
  /auth/login:
    post:
      tags: [认证]
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, example: scu123 }
                username: { type: string, description: 兼容旧客户端, deprecated: true }
                password: { type: string, example: "123456" }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenUser' }
...
  /auth/profile:
    put:
      tags: [认证]
      summary: 更新个人信息
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string, nullable: true, example: 新昵称 }
                bio: { type: string, nullable: true, example: 新的简介 }
                avatar_url: { type: string, nullable: true, example: "https://.../a.png" }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
  /auth/contact/send-code:
    post:
      tags: [认证]
      summary: 发送手机号或邮箱绑定验证码
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactTarget' }
      responses:
        '200':
          description: 发送成功
  /auth/contact/bind:
    post:
      tags: [认证]
      summary: 校验验证码并绑定手机号或邮箱
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ContactTarget'
                - type: object
                  required: [code]
                  properties:
                    code: { type: string, example: "123456" }
      responses:
        '200':
          description: 绑定成功
  /auth/password/change:
    post:
      tags: [认证]
      summary: 修改当前用户密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, example: "123456" }
                newPassword: { type: string, example: "654321" }
      responses:
        '200':
          description: 修改成功，返回新 token
  /auth/password/reset/send-code:
    post:
      tags: [认证]
      summary: 发送找回密码验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactTarget' }
      responses:
        '200':
          description: 请求已受理
  /auth/password/reset:
    post:
      tags: [认证]
      summary: 使用验证码重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ContactTarget'
                - type: object
                  required: [code, newPassword]
                  properties:
                    code: { type: string, example: "123456" }
                    newPassword: { type: string, example: "654321" }
      responses:
        '200':
          description: 密码重置成功，返回 token
  /users/{id}:
    get:
      tags: [认证]
      summary: 查看某用户公开信息
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }

  # ======================= 分类 + 标签 =======================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '登录后拿到的 token，放到请求头 "Authorization: Bearer <token>"'
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer, description: 0=成功, example: 0 }
        message: { type: string, example: success }
        data: {}
...
    TokenUser:
      type: object
      properties:
        token: { type: string, example: eyJhbGciOiJIUzI1NiIs... }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        nickname: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        role: { type: string, example: user }
        phone: { type: string, nullable: true, example: "138****8000" }
        email: { type: string, nullable: true, example: "u***@example.com" }
        phoneBound: { type: boolean }
        emailBound: { type: boolean }
    ContactTarget:
      type: object
      required: [channel, target]
      properties:
        channel: { type: string, enum: [phone, email] }
        target: { type: string, example: 13800138000 }
...